requirements.txt
python-dotenv
langgraph
langchain-groq
langchain-core
langchain-huggingface
numpy 
pandas 
mcp-use
